[FLINK-39113][s3] Fix missing encryption context in NativeS3OutputStream write path
Move serializeEncryptionContext to S3EncryptionConfig and apply it in
NativeS3OutputStream.uploadToS3(), which previously dropped the KMS
encryption context for small-file (non-multipart) writes.
